refactor(api): split usage schemas and routes into sibling modules
Move the 5 usage-tracking response schemas and the 2 usage endpoints
(plus their helpers) out of routes.py/schemas.py into dedicated
schemas_usage.py and routes_usage.py modules. The new router is
mounted alongside the main router in create_app and in the test
fixture. No behavior change; routes.py shrinks by 186 lines and
schemas.py by 112 lines.
